+++ changelog.md
@@ -1,4 +1,6 @@
 # Changelog
 
-**./test-parts/test-pathfinding-input.js - handlePossibleRoutes**
-* Check "Impossible Route" result. Must be zero.
+**./test-parts/test-pathfinding-input.js**
+* Wrote new function 'disposeExampleGraph'
+	* Disposes 'exampleGraph' after unit tests complete.
+	* Called from 'runTests'